+++ packages/mutation-testing-elements/test/integration/po/ResultTableRow.po.ts
@@ -15,9 +15,10 @@ export class ResultTableRow extends PageObject {
   public survived = () => this.$('td:nth-child(6)').getText();
   public timeout = () => this.$('td:nth-child(7)').getText();
   public noCoverage = () => this.$('td:nth-child(8)').getText();
-  public runtimeErrors = () => this.$('td:nth-child(9)').getText();
-  public compileErrors = () => this.$('td:nth-child(10)').getText();
-  public totalDetected = () => this.$('th:nth-child(11)').getText();
-  public totalUndetected = () => this.$('th:nth-child(12)').getText();
-  public totalMutants = () => this.$('th:nth-child(13)').getText();
+  public ignored = () => this.$('td:nth-child(9)').getText();
+  public runtimeErrors = () => this.$('td:nth-child(10)').getText();
+  public compileErrors = () => this.$('td:nth-child(11)').getText();
+  public totalDetected = () => this.$('th:nth-child(12)').getText();
+  public totalUndetected = () => this.$('th:nth-child(13)').getText();
+  public totalMutants = () => this.$('th:nth-child(14)').getText();
 }