+++ packages/mutation-testing-elements/test/unit/components/mutation-test-report-file-legend.spec.ts
@@ -9,7 +9,7 @@ describe(MutationTestReportFileLegendComponent.name, () => {
 
   beforeEach(async () => {
     sut = new CustomElementFixture('mutation-test-report-file-legend');
-    await sut.updateComplete;
+    await sut.whenStable();
   });
 
   afterEach(() => {
@@ -28,16 +28,18 @@ describe(MutationTestReportFileLegendComponent.name, () => {
         { status: MutantStatus.NoCoverage },
         { status: MutantStatus.RuntimeError },
         { status: MutantStatus.Survived },
-        { status: MutantStatus.Timeout }
+        { status: MutantStatus.Timeout },
+        { status: MutantStatus.Ignored }
       ];
-      await sut.updateComplete;
+      await sut.whenStable();
       const actualCheckboxes = sut.$$('.form-check.form-check-inline');
-      expect(actualCheckboxes).lengthOf(6);
-      const checkboxTexts = actualCheckboxes.map(checkbox => normalizeWhitespace(checkbox.textContent as string));
+      expect(actualCheckboxes).lengthOf(7);
+      const checkboxTexts = actualCheckboxes.map(checkbox => normalizeWhitespace((checkbox.textContent as string)));
       expect(checkboxTexts).deep.eq([
         '✅ Killed (1)',
         '👽 Survived (1)',
         '🙈 NoCoverage (1)',
+        '🤥 Ignored (1)',
         '⌛ Timeout (1)',
         '💥 CompileError (1)',
         '💥 RuntimeError (1)'
@@ -52,7 +54,7 @@ describe(MutationTestReportFileLegendComponent.name, () => {
         sut.element.mutants = [{ status: mutantStatus }];
 
         // Act
-        await sut.updateComplete;
+        await sut.whenStable();
 
         // Assert
         const badge = sut.$(`[data-status=${status}] .badge`);
@@ -69,25 +71,30 @@ describe(MutationTestReportFileLegendComponent.name, () => {
         { status: MutantStatus.NoCoverage },
         { status: MutantStatus.RuntimeError },
         { status: MutantStatus.Survived },
-        { status: MutantStatus.Timeout }
+        { status: MutantStatus.Timeout },
+        { status: MutantStatus.Ignored }
       ];
       const expected: MutantFilter[] = [
         { enabled: false, numberOfMutants: 1, status: MutantStatus.Killed },
         { enabled: true, numberOfMutants: 1, status: MutantStatus.Survived },
         { enabled: true, numberOfMutants: 1, status: MutantStatus.NoCoverage },
+        { enabled: false, numberOfMutants: 1, status: MutantStatus.Ignored },
         { enabled: true, numberOfMutants: 1, status: MutantStatus.Timeout },
         { enabled: false, numberOfMutants: 1, status: MutantStatus.CompileError },
-        { enabled: false, numberOfMutants: 1, status: MutantStatus.RuntimeError }
-      ];
-      await sut.updateComplete;
+        { enabled: false, numberOfMutants: 1, status: MutantStatus.RuntimeError }];
+      await sut.whenStable();
       expect(actualEvent).ok;
       expect((actualEvent as CustomEvent).detail).deep.eq(expected);
     });
 
     it('should dispatch the "filters-changed" event when a checkbox is checked', async () => {
       // Arrange
-      sut.element.mutants = [{ status: MutantStatus.CompileError }, { status: MutantStatus.Survived }];
-      await sut.updateComplete;
+      sut.element.mutants = [
+        { status: MutantStatus.CompileError },
+        { status: MutantStatus.Survived }
+      ];
+      const r = await sut.whenStable();
+      console.log(r);
       let actualEvent: CustomEvent | undefined;
       sut.element.addEventListener('filters-changed', (ev: any) => (actualEvent = ev));
       const expected: MutantFilter[] = [
@@ -97,7 +104,7 @@ describe(MutationTestReportFileLegendComponent.name, () => {
 
       // Act
       sut.$(`input[type="checkbox"][value="${MutantStatus.Survived}"]`).click();
-      await sut.updateComplete;
+      await sut.whenStable();
 
       // Assert
       expect(actualEvent).ok;
@@ -119,7 +126,7 @@ describe(MutationTestReportFileLegendComponent.name, () => {
       let actual: Event | undefined;
       sut.element.addEventListener('expand-all', evt => (actual = evt));
       collapseButton.click();
-      await sut.updateComplete;
+      await sut.whenStable();
       expect(actual).ok;
     });
 
@@ -127,16 +134,16 @@ describe(MutationTestReportFileLegendComponent.name, () => {
       let actual: Event | undefined;
       sut.element.addEventListener('collapse-all', evt => (actual = evt));
       collapseButton.click();
-      await sut.updateComplete;
+      await sut.whenStable();
       expect(actual).not.ok;
       collapseButton.click();
-      await sut.updateComplete;
+      await sut.whenStable();
       expect(actual).ok;
     });
 
     it('should toggle the text to "Collapse all" when clicked', async () => {
       collapseButton.click();
-      await sut.updateComplete;
+      await sut.whenStable();
       expect(collapseButton.textContent).eq('Collapse all');
     });
   });