+++ packages/mutation-testing-elements/test/unit/components/mutation-test-report-totals.spec.ts
@@ -8,7 +8,7 @@ describe(MutationTestReportTotalsComponent.name, () => {
 
   beforeEach(async () => {
     sut = new CustomElementFixture('mutation-test-report-totals');
-    await sut.updateComplete;
+    await sut.whenStable();
   });
 
   afterEach(() => {
@@ -23,11 +23,11 @@ describe(MutationTestReportTotalsComponent.name, () => {
     sut.element.model = createMetricsResult({
       file: createFileResult()
     });
-    await sut.updateComplete;
+    await sut.whenStable();
     const table = sut.$('table') as HTMLTableElement;
     expect(table).ok;
-    expect(table.querySelectorAll('thead th')).lengthOf(11);
-    expect(table.querySelectorAll('tbody th, tbody td')).lengthOf(13);
+    expect(table.querySelectorAll('thead th')).lengthOf(12);
+    expect(table.querySelectorAll('tbody th, tbody td')).lengthOf(14);
   });
 
   it('should show a table with a 3 rows for a directory result with 2 directories and one file', async () => {
@@ -39,7 +39,7 @@ describe(MutationTestReportTotalsComponent.name, () => {
       name: 'bar',
       childResults: [file, createMetricsResult({ name: 'baz' })]
     });
-    await sut.updateComplete;
+    await sut.whenStable();
     const table = sut.$('table') as HTMLTableElement;
     expect(table).ok;
     expect(table.querySelectorAll('tbody tr')).lengthOf(3);
@@ -62,7 +62,7 @@ describe(MutationTestReportTotalsComponent.name, () => {
     });
 
     // Act
-    await sut.updateComplete;
+    await sut.whenStable();
 
     // Assert
     const table = sut.$('table') as HTMLTableElement;
@@ -79,7 +79,7 @@ describe(MutationTestReportTotalsComponent.name, () => {
 
     sut.element.model.metrics.mutationScore = NaN;
 
-    await sut.updateComplete;
+    await sut.whenStable();
     const table = sut.$('table') as HTMLTableElement;
     expect(table).ok;
     expect(table.querySelectorAll('td span.font-weight-bold')[0].textContent).contains('N/A');
@@ -93,7 +93,7 @@ describe(MutationTestReportTotalsComponent.name, () => {
 
     sut.element.model.metrics.mutationScore = mutationScore;
 
-    await sut.updateComplete;
+    await sut.whenStable();
     const table = sut.$('table') as HTMLTableElement;
     expect(table).ok;
     expect(table.querySelectorAll('.progress')[0].textContent).contains(mutationScore);
@@ -106,7 +106,7 @@ describe(MutationTestReportTotalsComponent.name, () => {
 
     sut.element.model.metrics.mutationScore = NaN;
 
-    await sut.updateComplete;
+    await sut.whenStable();
     const table = sut.$('table') as HTMLTableElement;
     expect(table).ok;
     expect(table.querySelector('.progress')).null;