+++ packages/mutation-testing-metrics/src/api/Metrics.ts
@@ -29,6 +29,10 @@ export interface Metrics {
    */
   compileErrors: number;
   /**
+   * The total number of mutants that were not even tested because the config of the user asked for them to be ignored.
+   */
+  ignored: number;
+  /**
    * The total number of mutants that were detected, meaning either killed or caused a time out.
    * `killed + timed out`
    */
@@ -55,7 +59,7 @@ export interface Metrics {
   totalMutants: number;
   /**
    * The total number of mutants tested in an area that had code coverage result
-   * `totalDetected + survived`
+   * `totalDetected + survived + ignored`
    */
   totalCovered: number;
   /**