+++ packages/mutation-testing-metrics/src/calculateMetrics.ts
@@ -60,6 +60,7 @@ function countMetrics(mutants: Pick<MutantResult, 'status'>[]): Metrics {
   const noCoverage = count(MutantStatus.NoCoverage);
   const runtimeErrors = count(MutantStatus.RuntimeError);
   const compileErrors = count(MutantStatus.CompileError);
+  const ignored = count(MutantStatus.Ignored);
   const totalDetected = timeout + killed;
   const totalUndetected = survived + noCoverage;
   const totalCovered = totalDetected + survived;
@@ -72,13 +73,14 @@ function countMetrics(mutants: Pick<MutantResult, 'status'>[]): Metrics {
     noCoverage,
     runtimeErrors,
     compileErrors,
+    ignored,
     totalDetected,
     totalUndetected,
     totalCovered,
     totalValid,
     totalInvalid,
-    mutationScore: totalValid > 0 ? (totalDetected / totalValid) * 100 : DEFAULT_SCORE,
-    totalMutants: totalValid + totalInvalid,
-    mutationScoreBasedOnCoveredCode: totalValid > 0 ? (totalDetected / totalCovered) * 100 || 0 : DEFAULT_SCORE
+    mutationScore: totalValid > 0 ? totalDetected / totalValid * 100 : DEFAULT_SCORE,
+    totalMutants: totalValid + totalInvalid + ignored,
+    mutationScoreBasedOnCoveredCode: totalValid > 0 ? totalDetected / totalCovered * 100 || 0 : DEFAULT_SCORE
   };
 }