+++ packages/mutation-testing-elements/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.3...v1.3.0) (2020-02-27)
+
+
+### Bug Fixes
+
+* font styling being excluded by postcss ([#311](https://github.com/stryker-mutator/mutation-testing-elements/issues/311)) ([5e53bd1](https://github.com/stryker-mutator/mutation-testing-elements/commit/5e53bd11395f933c334f5a952db131645887c13a))
+
+
+### Features
+
+* **icons:** add icons to a few file types ([#173](https://github.com/stryker-mutator/mutation-testing-elements/issues/173)) ([a8752b8](https://github.com/stryker-mutator/mutation-testing-elements/commit/a8752b86ab956e4d96bd680e0c221c429174ea06))
+* **ignored status:** add MutantStatus Ignored ([#239](https://github.com/stryker-mutator/mutation-testing-elements/issues/239)) ([68b2302](https://github.com/stryker-mutator/mutation-testing-elements/commit/68b23022d2c4d21d642edc17ef4905c77adffd35)), closes [#85](https://github.com/stryker-mutator/mutation-testing-elements/issues/85)
+* **line-numbers:** add line numbers in code samples ([#313](https://github.com/stryker-mutator/mutation-testing-elements/issues/313)) ([20b3eab](https://github.com/stryker-mutator/mutation-testing-elements/commit/20b3eabfb89e9791d7425531736a17964f35d93e))
+* **n/a mutation score:** support no mutation score ([66a1549](https://github.com/stryker-mutator/mutation-testing-elements/commit/66a1549d7a464ccb42bdd988fa66a20e44430e43))
+
+
+
+
+
 ## [1.2.3](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.2...v1.2.3) (2019-12-11)
 
 **Note:** Version bump only for package mutation-testing-elements