+++ packages/mutation-testing-metrics-scala/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.3...v1.3.0) (2020-02-27)
+
+
+### Bug Fixes
+
+* **metrics-scala:** score should be NaN when there are 0 mutants ([#303](https://github.com/stryker-mutator/mutation-testing-elements/issues/303)) ([b3627e4](https://github.com/stryker-mutator/mutation-testing-elements/commit/b3627e427b8e0eb4bb13f7c5c6a6ee29caa99e44))
+
+
+### Features
+
+* **ignored status:** add MutantStatus Ignored ([#239](https://github.com/stryker-mutator/mutation-testing-elements/issues/239)) ([68b2302](https://github.com/stryker-mutator/mutation-testing-elements/commit/68b23022d2c4d21d642edc17ef4905c77adffd35)), closes [#85](https://github.com/stryker-mutator/mutation-testing-elements/issues/85)
+* **line-numbers:** add line numbers in code samples ([#313](https://github.com/stryker-mutator/mutation-testing-elements/issues/313)) ([20b3eab](https://github.com/stryker-mutator/mutation-testing-elements/commit/20b3eabfb89e9791d7425531736a17964f35d93e))
+
+
+
+
+
 ## [1.2.3](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.2...v1.2.3) (2019-12-11)
 
 **Note:** Version bump only for package mutation-testing-metrics-scala